NamedParameterJdbcTemplate

+++ Sql ~ Jpa/project/demo/demo/src/main/java/com/example/demo/repository/RoleDao.java
@@ -1,9 +1,12 @@
 package com.example.demo.repository;
 
 import com.example.demo.domain.Role;
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsertOperations;
@@ -17,16 +20,16 @@ import java.util.List;
 
 @Repository
 public class RoleDao {
-    private final JdbcTemplate jdbcTemplate; // 필드를 final로 선언하면 반드시 생성자에서 초기화
+    private final NamedParameterJdbcTemplate jdbcTemplate; // 필드를 final로 선언하면 반드시 생성자에서 초기화
     private SimpleJdbcInsertOperations insertAction; // insert를 쉽게 하도록 도와주는 인터페이스
 
     // 생성자에 파라미터를 넣어주면 스프링 부트가 자동으로 주입한다. 생성자 주입
     public RoleDao(DataSource dataSource) {
         System.out.println("RoleDao 생성자 호출");
         System.out.println(dataSource.getClass().getName());
-        this.jdbcTemplate = new JdbcTemplate(dataSource); // DataSource를 넣어줘야 한다
+        jdbcTemplate = new NamedParameterJdbcTemplate(dataSource); // DataSource를 넣어줘야 한다
 
-        this.insertAction = new SimpleJdbcInsert(dataSource)
+        insertAction = new SimpleJdbcInsert(dataSource)
                 .withTableName("role");
     }
 
@@ -46,24 +49,28 @@ public class RoleDao {
     }
 
     public boolean deleteRole(int roleId) {
-        String sql = "delete from role where role_id = ?";
-        int result = jdbcTemplate.update(sql, roleId);
+//        String sql = "delete from role where role_id = ?";
+//        int result = jdbcTemplate.update(sql, roleId);
+//        return result == 1;
+        String sql = "delete from role where role_id = :roleID";
+        SqlParameterSource params = new MapSqlParameterSource("roleId", roleId);
+        int result = jdbcTemplate.update(sql, params);
         return result == 1;
     }
 
     public Role getRole(int roleId) {
-        String sql = "select role_id, name from role where role_id = ?";
+        String sql = "select role_id, name from role where role_id = :roleID";
 
         // queryForObject는 1건 또는 0건을 읽어오는 메서드
         // queryForObject(String sql, RowMapper<T> rowMapper, @Nullable Object... args)
 
-
         try {
-            return jdbcTemplate.queryForObject(sql, new RoleRowMapper(), roleId);
+            SqlParameterSource params = new MapSqlParameterSource("roleId", roleId);
+            RowMapper<Role> roleRowMapper = BeanPropertyRowMapper.newInstance(Role.class);
+            return jdbcTemplate.queryForObject(sql, params, roleRowMapper); // 여기는 로우매퍼가 세 번째 파라미터로 들어감
         } catch (Exception e) {
             return null;
         }
-
     }
 
     // 위의 람다식을 풀어서 쓴 식
@@ -77,26 +84,22 @@ public class RoleDao {
 //        }, roleId);
 //    }
 
-    class RoleRowMapper implements RowMapper<Role> {
-        @Override
-        public Role mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Role role = new Role();
-            role.setRoleId(rs.getInt("role_id"));
-            role.setName(rs.getString("name"));
-            return role;
-        }
-
-    }
+//    class RoleRowMapper implements RowMapper<Role> {
+//        @Override
+//        public Role mapRow(ResultSet rs, int rowNum) throws SQLException {
+//            Role role = new Role();
+//            role.setRoleId(rs.getInt("role_id"));
+//            role.setName(rs.getString("name"));
+//            return role;
+//        }
+//
+//    }
 
     public List<Role> getRoles() {
         String sql = "select role_id, name from role order by role_id";
         // query 메서드는 여러 건의 결과를 구할때 사용
-        return jdbcTemplate.query(sql, (rs, rowNum) -> {
-            Role role = new Role();
-            role.setRoleId(rs.getInt("role_id"));
-            role.setName(rs.getString("name"));
-            return role;
-        });
+        RowMapper<Role> roleRowMapper = BeanPropertyRowMapper.newInstance(Role.class);
+        return jdbcTemplate.query(sql, roleRowMapper);
     }
 }
